requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
numpy==2.3.3
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
